refactor: group relaycast broker integration
