+++ .trajectories/index.json
@@ -1,6 +1,6 @@
 {
   "version": 1,
-  "lastUpdated": "2026-05-19T02:56:24.778Z",
+  "lastUpdated": "2026-05-19T03:54:07.051Z",
   "trajectories": {
     "traj_05xg7j388bc4": {
       "title": "Add browser workflow step integration",
@@ -988,6 +988,13 @@
       "startedAt": "2026-05-19T02:48:10.768Z",
       "completedAt": "2026-05-19T02:56:24.584Z",
       "path": "/Users/will/Projects/AgentWorkforce/relay/.trajectories/completed/2026-05/traj_jbo2x14y7ovt.json"
+    },
+    "traj_5qbla7w4kzoi": {
+      "title": "Fix issue 877",
+      "status": "completed",
+      "startedAt": "2026-05-19T03:40:40.798Z",
+      "completedAt": "2026-05-19T03:54:06.889Z",
+      "path": "/Users/will/Projects/AgentWorkforce/relay/.trajectories/completed/2026-05/traj_5qbla7w4kzoi.json"
     }
   }
 }
