+++ crates/broker/src/relaycast/auth.rs
@@ -1,0 +1,1304 @@
+use anyhow::{Context, Result};
+use chrono::{DateTime, Utc};
+use relaycast::{CreateAgentRequest, RelayCast, RelayCastOptions, RelayError};
+use reqwest::StatusCode;
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+use sha2::{Digest, Sha256};
+use uuid::Uuid;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct WorkspaceCredential {
+    pub workspace_id: String,
+    #[serde(
+        default,
+        alias = "workspaceAlias",
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub workspace_alias: Option<String>,
+    pub agent_id: String,
+    pub api_key: String,
+    #[serde(default)]
+    pub agent_name: Option<String>,
+    #[serde(default)]
+    pub agent_token: Option<String>,
+    pub updated_at: DateTime<Utc>,
+}
+
+pub type CredentialCache = WorkspaceCredential;
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+pub struct CredentialSet {
+    #[serde(default)]
+    pub memberships: Vec<WorkspaceCredential>,
+    #[serde(
+        default,
+        alias = "defaultWorkspaceId",
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub default_workspace_id: Option<String>,
+}
+
+#[derive(Debug, Clone)]
+pub struct AuthSessionSet {
+    pub memberships: Vec<AuthSession>,
+    pub default_workspace_id: Option<String>,
+}
+
+#[derive(Debug, Clone)]
+pub struct AuthSession {
+    pub credentials: CredentialCache,
+    pub token: String,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+struct WorkspaceSource {
+    #[serde(default, alias = "workspaceId")]
+    workspace_id: Option<String>,
+    #[serde(default, alias = "workspaceAlias")]
+    workspace_alias: Option<String>,
+    api_key: String,
+}
+
+impl CredentialSet {
+    pub fn from_json(raw: &str) -> Result<Self> {
+        let value: Value = serde_json::from_str(raw).context("invalid credential set JSON")?;
+        Self::from_value(value)
+    }
+
+    pub fn from_value(value: Value) -> Result<Self> {
+        if let Ok(set) = serde_json::from_value::<CredentialSet>(value.clone()) {
+            if !set.memberships.is_empty() {
+                return Ok(Self::normalize(set));
+            }
+        }
+
+        if let Ok(legacy) = serde_json::from_value::<WorkspaceCredential>(value.clone()) {
+            return Ok(Self::from_legacy(legacy));
+        }
+
+        if let Ok(legacy) = serde_json::from_value::<Vec<WorkspaceCredential>>(value.clone()) {
+            return Ok(Self::from_memberships(legacy, None));
+        }
+
+        if let Ok(source) = serde_json::from_value::<WorkspaceSource>(value.clone()) {
+            return Ok(Self::from_memberships(
+                vec![WorkspaceCredential {
+                    workspace_id: source
+                        .workspace_id
+                        .unwrap_or_else(|| "ws_unknown".to_string()),
+                    workspace_alias: source.workspace_alias,
+                    agent_id: String::new(),
+                    api_key: source.api_key,
+                    agent_name: None,
+                    agent_token: None,
+                    updated_at: Utc::now(),
+                }],
+                None,
+            ));
+        }
+
+        anyhow::bail!("credential JSON was neither a credential set nor a legacy cache entry")
+    }
+
+    pub fn from_legacy(legacy: WorkspaceCredential) -> Self {
+        let default_workspace_id = Some(legacy.workspace_id.clone());
+        Self {
+            memberships: vec![legacy],
+            default_workspace_id,
+        }
+    }
+
+    pub fn from_memberships(
+        memberships: Vec<WorkspaceCredential>,
+        default_workspace_id: Option<String>,
+    ) -> Self {
+        Self::normalize(Self {
+            memberships,
+            default_workspace_id,
+        })
+    }
+
+    pub fn default_membership(&self) -> Option<&WorkspaceCredential> {
+        if let Some(default_workspace_id) = self.default_workspace_id.as_deref() {
+            self.membership_by_selector(default_workspace_id)
+                .or_else(|| {
+                    self.memberships
+                        .iter()
+                        .find(|membership| membership.workspace_id == default_workspace_id)
+                })
+        } else if self.memberships.len() == 1 {
+            self.memberships.first()
+        } else {
+            None
+        }
+    }
+
+    pub fn membership_by_selector(&self, selector: &str) -> Option<&WorkspaceCredential> {
+        let trimmed = selector.trim();
+        self.memberships.iter().find(|membership| {
+            membership.workspace_id == trimmed
+                || membership
+                    .workspace_alias
+                    .as_deref()
+                    .is_some_and(|alias| alias.eq_ignore_ascii_case(trimmed))
+        })
+    }
+
+    fn normalize(mut set: Self) -> Self {
+        set.memberships
+            .retain(|membership| !membership.api_key.trim().is_empty());
+        if set.default_workspace_id.is_none() && set.memberships.len() == 1 {
+            set.default_workspace_id = set
+                .memberships
+                .first()
+                .map(|membership| membership.workspace_id.clone());
+        }
+        set
+    }
+}
+
+impl AuthSessionSet {
+    pub fn credential_set(&self) -> CredentialSet {
+        CredentialSet::from_memberships(
+            self.memberships
+                .iter()
+                .map(|session| session.credentials.clone())
+                .collect(),
+            self.default_workspace_id.clone(),
+        )
+    }
+
+    pub fn default_session(&self) -> Option<&AuthSession> {
+        if let Some(default_workspace_id) = self.default_workspace_id.as_deref() {
+            self.memberships.iter().find(|session| {
+                session.credentials.workspace_id == default_workspace_id
+                    || session
+                        .credentials
+                        .workspace_alias
+                        .as_deref()
+                        .is_some_and(|alias| alias.eq_ignore_ascii_case(default_workspace_id))
+            })
+        } else if self.memberships.len() == 1 {
+            self.memberships.first()
+        } else {
+            None
+        }
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+#[error("{message}")]
+struct AuthHttpError {
+    status: StatusCode,
+    message: String,
+}
+
+/// Generate a deterministic workspace name based on the current user and working
+/// directory so that the same user in the same project gets the same workspace
+/// across sessions (enabling message continuity and resume).
+fn deterministic_workspace_name() -> String {
+    let user = std::env::var("USER")
+        .or_else(|_| std::env::var("USERNAME"))
+        .unwrap_or_else(|_| "unknown".to_string());
+    let cwd = std::env::current_dir()
+        .map(|p| p.to_string_lossy().to_string())
+        .unwrap_or_else(|_| "unknown".to_string());
+
+    let mut hasher = Sha256::new();
+    hasher.update(format!("{user}:{cwd}"));
+    let hash = format!("{:x}", hasher.finalize());
+    format!("relay-{}", &hash[..8])
+}
+
+#[derive(Clone)]
+pub struct AuthClient {
+    base_url: String,
+}
+
+impl AuthClient {
+    pub fn new(base_url: impl Into<String>) -> Self {
+        Self {
+            base_url: base_url.into(),
+        }
+    }
+
+    pub async fn startup_session(&self, requested_name: Option<&str>) -> Result<AuthSession> {
+        self.startup_session_with_options(requested_name, false, None)
+            .await
+    }
+
+    pub async fn startup_session_set(
+        &self,
+        requested_name: Option<&str>,
+    ) -> Result<AuthSessionSet> {
+        self.startup_session_set_with_options(requested_name, false, None)
+            .await
+    }
+
+    pub async fn startup_session_with_options(
+        &self,
+        requested_name: Option<&str>,
+        strict_name: bool,
+        agent_type: Option<&str>,
+    ) -> Result<AuthSession> {
+        let sessions = self
+            .startup_session_set_with_options(requested_name, strict_name, agent_type)
+            .await?;
+        sessions
+            .default_session()
+            .cloned()
+            .context("no default workspace session was available")
+    }
+
+    pub async fn startup_session_set_with_options(
+        &self,
+        requested_name: Option<&str>,
+        strict_name: bool,
+        agent_type: Option<&str>,
+    ) -> Result<AuthSessionSet> {
+        if let Some((sources, default_hint)) = self.load_workspace_sources_from_env()? {
+            let preferred_name = requested_name;
+            let mut memberships = Vec::with_capacity(sources.len());
+            let mut auth_rejections = Vec::new();
+
+            for source in sources {
+                let Some(api_key) = normalize_workspace_key(&source.api_key) else {
+                    anyhow::bail!("RELAY_WORKSPACES_JSON contained an invalid workspace key");
+                };
+                match self
+                    .register_agent_with_workspace_key(
+                        &api_key,
+                        preferred_name,
+                        strict_name,
+                        agent_type,
+                    )
+                    .await
+                {
+                    Ok(registration) => {
+                        let mut session = self.finish_session(
+                            api_key,
+                            source.workspace_id.clone(),
+                            registration,
+                        )?;
+                        session.credentials.workspace_alias = source.workspace_alias.clone();
+                        memberships.push(session);
+                    }
+                    Err(error) if is_auth_rejection(&error) => {
+                        auth_rejections
+                            .push(source.workspace_id.unwrap_or_else(|| "env".to_string()));
+                    }
+                    Err(error) if is_rate_limited(&error) => {
+                        auth_rejections.push(
+                            source
+                                .workspace_id
+                                .unwrap_or_else(|| "env_rate_limited".to_string()),
+                        );
+                    }
+                    Err(error) => {
+                        return Err(error)
+                            .context("failed registering agent for configured workspace");
+                    }
+                }
+            }
+
+            if memberships.is_empty() {
+                anyhow::bail!(
+                    "all configured multi-workspace memberships were rejected ({})",
+                    auth_rejections.join(", ")
+                );
+            }
+
+            return Ok(AuthSessionSet {
+                default_workspace_id: resolve_default_workspace_id(
+                    default_hint.as_deref(),
+                    &memberships,
+                ),
+                memberships,
+            });
+        }
+
+        self.startup_single_session_set_from_sources(requested_name, strict_name, agent_type)
+            .await
+    }
+
+    /// Rotate the token for an existing agent without re-registering.
+    ///
+    /// Calls `POST /v1/agents/:name/rotate-token` which generates a new bearer
+    /// token while keeping the same agent identity and name. Falls back to full
+    /// `refresh_session` if the agent no longer exists (404).
+    pub async fn rotate_token(&self, cached: &CredentialCache) -> Result<AuthSession> {
+        match self.rotate_token_no_fallback(cached).await {
+            Ok(session) => Ok(session),
+            Err(error) if is_not_found(&error) => {
+                let agent_name = cached
+                    .agent_name
+                    .as_deref()
+                    .context("cannot rotate token without agent name")?;
+                let api_key = normalize_workspace_key(&cached.api_key)
+                    .context("cached api_key is not a valid workspace key")?;
+                tracing::info!(
+                    target = "relay_broker::auth",
+                    agent_name = %agent_name,
+                    "agent not found during token rotation, falling back to re-registration"
+                );
+                let registration = self
+                    .register_agent_with_workspace_key(&api_key, Some(agent_name), false, None)
+                    .await
+                    .context("failed to re-register after rotate-token 404")?;
+                let mut session =
+                    self.finish_session(api_key, Some(cached.workspace_id.clone()), registration)?;
+                session.credentials.workspace_alias = cached.workspace_alias.clone();
+                Ok(session)
+            }
+            Err(error) => Err(error),
+        }
+    }
+
+    async fn rotate_token_no_fallback(&self, cached: &CredentialCache) -> Result<AuthSession> {
+        let agent_name = cached
+            .agent_name
+            .as_deref()
+            .context("cannot rotate token without agent name")?;
+        let api_key = normalize_workspace_key(&cached.api_key)
+            .context("cached api_key is not a valid workspace key")?;
+
+        let relay = build_relay_client(&api_key, &self.base_url)?;
+        let result = relay
+            .rotate_agent_token(agent_name)
+            .await
+            .map_err(relay_error_to_anyhow)?;
+        let token = result.token;
+
+        let creds = CredentialCache {
+            workspace_id: cached.workspace_id.clone(),
+            workspace_alias: cached.workspace_alias.clone(),
+            agent_id: cached.agent_id.clone(),
+            api_key: cached.api_key.clone(),
+            agent_name: Some(agent_name.to_string()),
+            agent_token: Some(token.clone()),
+            updated_at: Utc::now(),
+        };
+
+        Ok(AuthSession {
+            credentials: creds,
+            token,
+        })
+    }
+
+    async fn startup_single_session_set_from_sources(
+        &self,
+        requested_name: Option<&str>,
+        strict_name: bool,
+        agent_type: Option<&str>,
+    ) -> Result<AuthSessionSet> {
+        let env_workspace_key = std::env::var("RELAY_API_KEY")
+            .ok()
+            .and_then(|s| normalize_workspace_key(&s));
+
+        let mut workspace_id_hint: Option<String> = None;
+
+        let mut candidates: Vec<(&str, String)> = Vec::new();
+        if let Some(key) = env_workspace_key {
+            candidates.push(("env", key));
+        }
+
+        let mut attempted_fresh_workspace = false;
+        if candidates.is_empty() {
+            let ws_name = deterministic_workspace_name();
+            let (workspace_id, api_key) = self.create_workspace(&ws_name).await?;
+            workspace_id_hint = Some(workspace_id);
+            candidates.push(("fresh", api_key));
+            attempted_fresh_workspace = true;
+        }
+
+        let preferred_name = requested_name;
+        let mut auth_rejections = Vec::new();
+
+        for (source, key) in &candidates {
+            tracing::info!(
+                target = "relay_broker::auth",
+                source = %source,
+                preferred_name = ?preferred_name,
+                strict_name = %strict_name,
+                agent_type = ?agent_type,
+                "attempting registration with workspace key"
+            );
+            match self
+                .register_agent_with_workspace_key(key, preferred_name, strict_name, agent_type)
+                .await
+            {
+                Ok(registration) => {
+                    tracing::info!(
+                        target = "relay_broker::auth",
+                        agent_id = %registration.0,
+                        returned_name = %registration.1,
+                        "registration succeeded"
+                    );
+                    let session =
+                        self.finish_session(key.clone(), workspace_id_hint.clone(), registration)?;
+                    return Ok(AuthSessionSet {
+                        default_workspace_id: Some(session.credentials.workspace_id.clone()),
+                        memberships: vec![session],
+                    });
+                }
+                Err(error) if is_auth_rejection(&error) => {
+                    auth_rejections.push(format!("{source} key rejected"));
+                }
+                Err(error) if is_rate_limited(&error) => {
+                    auth_rejections.push(format!("{source} key rate-limited"));
+                }
+                Err(error) => {
+                    return Err(error).context(format!(
+                        "failed registering agent with {source} workspace key"
+                    ));
+                }
+            }
+        }
+
+        if !attempted_fresh_workspace {
+            let ws_name = deterministic_workspace_name();
+            let (workspace_id, api_key) = self.create_workspace(&ws_name).await?;
+            workspace_id_hint = Some(workspace_id);
+            match self
+                .register_agent_with_workspace_key(
+                    &api_key,
+                    preferred_name,
+                    strict_name,
+                    agent_type,
+                )
+                .await
+            {
+                Ok(registration) => {
+                    let session = self.finish_session(api_key, workspace_id_hint, registration)?;
+                    return Ok(AuthSessionSet {
+                        default_workspace_id: Some(session.credentials.workspace_id.clone()),
+                        memberships: vec![session],
+                    });
+                }
+                Err(error) => {
+                    return Err(error).context("failed registering agent with fresh workspace key");
+                }
+            }
+        }
+
+        anyhow::bail!(
+            "all workspace keys were rejected ({})",
+            auth_rejections.join(", ")
+        );
+    }
+
+    fn load_workspace_sources_from_env(
+        &self,
+    ) -> Result<Option<(Vec<WorkspaceSource>, Option<String>)>> {
+        let Ok(raw) = std::env::var("RELAY_WORKSPACES_JSON") else {
+            return Ok(None);
+        };
+        let trimmed = raw.trim();
+        if trimmed.is_empty() {
+            return Ok(None);
+        }
+
+        let value: Value =
+            serde_json::from_str(trimmed).context("RELAY_WORKSPACES_JSON must be valid JSON")?;
+        let default_hint = std::env::var("RELAY_DEFAULT_WORKSPACE")
+            .ok()
+            .map(|value| value.trim().to_string())
+            .filter(|value| !value.is_empty())
+            .or_else(|| {
+                value
+                    .get("default_workspace_id")
+                    .or_else(|| value.get("defaultWorkspaceId"))
+                    .and_then(Value::as_str)
+                    .map(str::trim)
+                    .filter(|value| !value.is_empty())
+                    .map(str::to_string)
+            });
+
+        let membership_values = if let Some(arr) = value.as_array() {
+            arr.clone()
+        } else if let Some(arr) = value.get("memberships").and_then(Value::as_array).cloned() {
+            arr
+        } else {
+            vec![value]
+        };
+
+        let mut sources = Vec::with_capacity(membership_values.len());
+        for membership in membership_values {
+            if let Ok(source) = serde_json::from_value::<WorkspaceSource>(membership.clone()) {
+                sources.push(source);
+                continue;
+            }
+            if let Ok(credential) = serde_json::from_value::<WorkspaceCredential>(membership) {
+                sources.push(WorkspaceSource {
+                    workspace_id: Some(credential.workspace_id),
+                    workspace_alias: credential.workspace_alias,
+                    api_key: credential.api_key,
+                });
+                continue;
+            }
+            anyhow::bail!("RELAY_WORKSPACES_JSON contains an invalid membership entry");
+        }
+
+        if sources.is_empty() {
+            anyhow::bail!("RELAY_WORKSPACES_JSON must contain at least one membership");
+        }
+
+        Ok(Some((sources, default_hint)))
+    }
+
+    fn finish_session(
+        &self,
+        workspace_key: String,
+        workspace_id_hint: Option<String>,
+        registration: (String, String, String, Option<String>),
+    ) -> Result<AuthSession> {
+        let (agent_id, agent_name, token, workspace_id_from_register) = registration;
+        let workspace_id = workspace_id_from_register
+            .or(workspace_id_hint)
+            .unwrap_or_else(|| "ws_unknown".to_string());
+
+        let creds = CredentialCache {
+            workspace_id,
+            workspace_alias: None,
+            agent_id,
+            api_key: workspace_key,
+            agent_name: Some(agent_name),
+            agent_token: Some(token.clone()),
+            updated_at: Utc::now(),
+        };
+
+        Ok(AuthSession {
+            credentials: creds,
+            token,
+        })
+    }
+
+    async fn create_workspace(&self, name: &str) -> Result<(String, String)> {
+        match RelayCast::create_workspace(name, Some(&self.base_url)).await {
+            Ok(result) => Ok((result.workspace_id, result.api_key)),
+            Err(error) if is_workspace_name_conflict(&error) => {
+                let suffix = Uuid::new_v4().simple().to_string();
+                let fallback_name = format!("{name}-{}", &suffix[..8]);
+                tracing::warn!(
+                    workspace_name = %name,
+                    fallback_name = %fallback_name,
+                    "workspace already exists; retrying with a fresh fallback name"
+                );
+                let result = RelayCast::create_workspace(&fallback_name, Some(&self.base_url))
+                    .await
+                    .map_err(relay_error_to_anyhow)?;
+                Ok((result.workspace_id, result.api_key))
+            }
+            Err(error) => Err(relay_error_to_anyhow(error)),
+        }
+    }
+
+    async fn register_agent_with_workspace_key(
+        &self,
+        workspace_key: &str,
+        requested_name: Option<&str>,
+        strict_name: bool,
+        agent_type: Option<&str>,
+    ) -> Result<(String, String, String, Option<String>)> {
+        let relay = build_relay_client(workspace_key, &self.base_url)?;
+        let mut attempted_retry = false;
+        let mut name = requested_name
+            .map(ToOwned::to_owned)
+            .unwrap_or_else(|| format!("agent-{}", Uuid::new_v4().simple()));
+
+        if strict_name {
+            let request = CreateAgentRequest {
+                name,
+                agent_type: Some(agent_type.unwrap_or("agent").to_string()),
+                persona: None,
+                metadata: None,
+            };
+            let result = relay
+                .register_or_get_agent(request)
+                .await
+                .map_err(relay_error_to_anyhow)?;
+            return Ok((
+                result.id,
+                result.name,
+                result.token,
+                None, // workspace_id not returned in CreateAgentResponse
+            ));
+        }
+
+        loop {
+            let request = CreateAgentRequest {
+                name: name.clone(),
+                agent_type: Some(agent_type.unwrap_or("agent").to_string()),
+                persona: None,
+                metadata: None,
+            };
+
+            match relay.register_agent(request).await {
+                Ok(result) => {
+                    return Ok((
+                        result.id,
+                        result.name,
+                        result.token,
+                        None, // workspace_id not returned in CreateAgentResponse
+                    ));
+                }
+                Err(RelayError::Api { code, status, .. })
+                    if is_conflict_code(&code) || status == 409 =>
+                {
+                    if !attempted_retry {
+                        attempted_retry = true;
+                        let suffix = Uuid::new_v4().simple().to_string();
+                        name = format!("{}-{}", name, &suffix[..8]);
+                        continue;
+                    }
+                    return Err(relay_error_to_anyhow(RelayError::Api {
+                        code: "agent_already_exists".to_string(),
+                        message: format!("agent name '{}' already exists after retry", name),
+                        status: 409,
+                    }));
+                }
+                Err(error) => {
+                    return Err(relay_error_to_anyhow(error));
+                }
+            }
+        }
+    }
+
+    pub async fn workspace_key_is_live(&self, workspace_key: &str) -> Result<bool> {
+        let Some(workspace_key) = normalize_workspace_key(workspace_key) else {
+            return Ok(false);
+        };
+        let relay = match build_relay_client(&workspace_key, &self.base_url) {
+            Ok(relay) => relay,
+            Err(_) => return Ok(false),
+        };
+        match relay.list_channels(false).await {
+            Ok(_) => Ok(true),
+            Err(RelayError::Api { status, .. }) if status == 401 || status == 403 => Ok(false),
+            Err(_) => Ok(false),
+        }
+    }
+}
+
+fn resolve_default_workspace_id(
+    selector: Option<&str>,
+    memberships: &[AuthSession],
+) -> Option<String> {
+    if let Some(selector) = selector.map(str::trim).filter(|value| !value.is_empty()) {
+        return memberships.iter().find_map(|session| {
+            if session.credentials.workspace_id == selector
+                || session
+                    .credentials
+                    .workspace_alias
+                    .as_deref()
+                    .is_some_and(|alias| alias.eq_ignore_ascii_case(selector))
+            {
+                Some(session.credentials.workspace_id.clone())
+            } else {
+                None
+            }
+        });
+    }
+
+    if memberships.len() == 1 {
+        memberships
+            .first()
+            .map(|session| session.credentials.workspace_id.clone())
+    } else {
+        None
+    }
+}
+
+fn normalize_workspace_key(raw: &str) -> Option<String> {
+    let trimmed = raw.trim();
+    if trimmed.starts_with("rk_") {
+        Some(trimmed.to_string())
+    } else {
+        None
+    }
+}
+
+fn is_auth_rejection(err: &anyhow::Error) -> bool {
+    auth_http_status(err)
+        .is_some_and(|status| status == StatusCode::UNAUTHORIZED || status == StatusCode::FORBIDDEN)
+}
+
+fn is_not_found(err: &anyhow::Error) -> bool {
+    auth_http_status(err).is_some_and(|status| status == StatusCode::NOT_FOUND)
+}
+
+fn is_rate_limited(err: &anyhow::Error) -> bool {
+    auth_http_status(err).is_some_and(|status| status == StatusCode::TOO_MANY_REQUESTS)
+}
+
+fn auth_http_status(err: &anyhow::Error) -> Option<StatusCode> {
+    err.downcast_ref::<AuthHttpError>()
+        .map(|e| e.status)
+        .or_else(|| {
+            err.downcast_ref::<reqwest::Error>()
+                .and_then(reqwest::Error::status)
+        })
+}
+
+/// Build a `RelayCast` workspace client from an API key and base URL.
+fn build_relay_client(api_key: &str, base_url: &str) -> Result<RelayCast> {
+    let opts = RelayCastOptions::new(api_key).with_base_url(base_url);
+    RelayCast::new(opts).map_err(|e| anyhow::anyhow!("{e}"))
+}
+
+/// Convert a `RelayError` into an `anyhow::Error`, preserving the HTTP status
+/// so that `is_auth_rejection`, `is_not_found`, and `is_rate_limited` still work.
+fn relay_error_to_anyhow(error: RelayError) -> anyhow::Error {
+    match &error {
+        RelayError::Api {
+            status, message, ..
+        } => anyhow::Error::new(AuthHttpError {
+            status: StatusCode::from_u16(*status).unwrap_or(StatusCode::INTERNAL_SERVER_ERROR),
+            message: message.clone(),
+        }),
+        _ => anyhow::anyhow!("{error}"),
+    }
+}
+
+fn is_conflict_code(code: &str) -> bool {
+    matches!(
+        code,
+        "agent_already_exists" | "name_taken" | "conflict" | "duplicate"
+    )
+}
+
+fn is_workspace_name_conflict(error: &RelayError) -> bool {
+    match error {
+        RelayError::Api {
+            code,
+            message,
+            status,
+        } => {
+            *status == 409
+                || is_conflict_code(code)
+                || message.to_ascii_lowercase().contains("already exists")
+        }
+        _ => false,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    // These tests intentionally hold the process-env mutex across awaits so
+    // concurrent async tests cannot mutate RELAY_* variables underneath each
+    // other.
+    #![allow(clippy::await_holding_lock)]
+
+    use std::sync::{Mutex, MutexGuard};
+
+    use httpmock::Method::{GET, POST};
+    use httpmock::MockServer;
+    use serde_json::json;
+
+    use super::{AuthClient, CredentialCache};
+
+    static RELAY_ENV_MUTEX: Mutex<()> = Mutex::new(());
+
+    /// Remove RELAY_API_KEY from the environment so it doesn't interfere with
+    /// mock-server tests. Tests use httpmock and only set up specific auth
+    /// headers — the real env key causes 404s against the mock.
+    fn clear_relay_env() -> MutexGuard<'static, ()> {
+        let guard = RELAY_ENV_MUTEX.lock().unwrap();
+        // SAFETY: test-only; Rust warns about remove_var in multi-threaded
+        // contexts but we accept the risk in test code.
+        unsafe {
+            std::env::remove_var("RELAY_API_KEY");
+            std::env::remove_var("RELAY_WORKSPACES_JSON");
+            std::env::remove_var("RELAY_DEFAULT_WORKSPACE");
+        }
+        guard
+    }
+
+    #[tokio::test]
+    async fn first_run_creates_workspace_and_agent_session() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let workspace = server.mock(|when, then| {
+            when.method(POST).path("/v1/workspaces");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"workspace_id":"ws_new","api_key":"rk_live_new","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+        let register = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_new");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"id":"a1","name":"lead","token":"at_live_1","status":"online","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+
+        let client = AuthClient::new(server.base_url());
+
+        let session = client.startup_session(Some("lead")).await.unwrap();
+        assert_eq!(session.token, "at_live_1");
+        assert_eq!(session.credentials.api_key, "rk_live_new");
+        assert_eq!(session.credentials.workspace_id, "ws_new");
+        assert_eq!(session.credentials.agent_id, "a1");
+        assert_eq!(session.credentials.agent_name.as_deref(), Some("lead"));
+
+        workspace.assert_hits(1);
+        register.assert_hits(1);
+    }
+
+    #[tokio::test]
+    async fn uses_env_workspace_key_without_creating_workspace() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        unsafe {
+            std::env::set_var("RELAY_API_KEY", "rk_live_env");
+        }
+        let register = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_env");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"id":"a2","name":"lead","token":"at_live_2","status":"online","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+
+        let client = AuthClient::new(server.base_url());
+
+        let session = client.startup_session(Some("lead")).await.unwrap();
+        assert_eq!(session.token, "at_live_2");
+        assert_eq!(session.credentials.api_key, "rk_live_env");
+        register.assert_hits(1);
+
+        unsafe {
+            std::env::remove_var("RELAY_API_KEY");
+        }
+    }
+
+    #[tokio::test]
+    async fn unauthorized_env_key_falls_back_to_fresh_workspace() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let _stale_register = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_stale");
+            then.status(401)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":false,"error":{"code":"unauthorized","message":"unauthorized"}}"#);
+        });
+        let workspace = server.mock(|when, then| {
+            when.method(POST).path("/v1/workspaces");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"workspace_id":"ws_new","api_key":"rk_live_new","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+        let fresh_register = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_new");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"id":"a9","name":"lead","token":"at_live_9","status":"online","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+
+        unsafe {
+            std::env::set_var("RELAY_API_KEY", "rk_live_stale");
+        }
+
+        let client = AuthClient::new(server.base_url());
+        let session = client.startup_session(Some("lead")).await.unwrap();
+        assert_eq!(session.token, "at_live_9");
+        assert_eq!(session.credentials.api_key, "rk_live_new");
+        assert_eq!(session.credentials.workspace_id, "ws_new");
+        workspace.assert_hits(1);
+        fresh_register.assert_hits(1);
+
+        unsafe {
+            std::env::remove_var("RELAY_API_KEY");
+        }
+    }
+
+    #[tokio::test]
+    async fn strict_name_conflict_reclaims_via_sdk_register_or_get_agent() {
+        // Regression test for issue #797: when a broker is restarted (or a
+        // second broker joins via shared workspace key) with a name that's
+        // already registered, registration must reclaim the existing agent
+        // through the relaycast SDK instead of failing the broker startup.
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        unsafe {
+            std::env::set_var("RELAY_API_KEY", "rk_live_shared");
+        }
+        let conflict = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_shared")
+                .json_body(json!({
+                    "name": "lead",
+                    "type": "agent"
+                }));
+            then.status(409)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":false,"error":{"code":"agent_already_exists","message":"name_taken"}}"#);
+        });
+        let get_existing = server.mock(|when, then| {
+            when.method(GET)
+                .path("/v1/agents/lead")
+                .header("authorization", "Bearer rk_live_shared");
+            then.status(200)
+                .header("content-type", "application/json")
+                // Mirrors the live cloud's GET /v1/agents/{name} payload that
+                // relaycast 1.0.1 accepts while reclaiming the agent.
+                .body(r#"{"ok":true,"data":{"id":"a_existing","name":"lead","type":"agent","status":"offline","persona":null,"metadata":{},"last_seen":"2025-01-01T00:00:00Z","channels":[]}}"#);
+        });
+        let rotate = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents/lead/rotate-token")
+                .header("authorization", "Bearer rk_live_shared");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"name":"lead","token":"at_live_rotated"}}"#);
+        });
+
+        let client = AuthClient::new(server.base_url());
+        let session = client
+            .startup_session_with_options(Some("lead"), true, None)
+            .await
+            .expect("strict-name conflict should reclaim via relaycast SDK");
+
+        assert_eq!(session.token, "at_live_rotated");
+        assert_eq!(session.credentials.agent_id, "a_existing");
+        assert_eq!(session.credentials.api_key, "rk_live_shared");
+        assert_eq!(session.credentials.agent_name.as_deref(), Some("lead"));
+        conflict.assert_hits(1);
+        get_existing.assert_hits(1);
+        rotate.assert_hits(1);
+
+        unsafe {
+            std::env::remove_var("RELAY_API_KEY");
+        }
+    }
+
+    #[tokio::test]
+    async fn default_name_conflict_retries_with_suffix_once() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let workspace = server.mock(|when, then| {
+            when.method(POST).path("/v1/workspaces");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"workspace_id":"ws_new","api_key":"rk_live_cached","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+        let first_conflict = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_cached")
+                .json_body(json!({
+                    "name": "lead",
+                    "type": "agent"
+                }));
+            then.status(409)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":false,"error":{"code":"agent_already_exists","message":"name_taken"}}"#);
+        });
+        let second_success = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_cached")
+                .body_contains("\"name\":\"lead-");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"id":"a10","name":"lead-suffixed","token":"at_live_10","status":"online","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+
+        let client = AuthClient::new(server.base_url());
+        let session = client.startup_session(Some("lead")).await.unwrap();
+
+        assert_eq!(session.token, "at_live_10");
+        assert_eq!(
+            session.credentials.agent_name.as_deref(),
+            Some("lead-suffixed")
+        );
+        workspace.assert_hits(1);
+        first_conflict.assert_hits(1);
+        second_success.assert_hits(1);
+    }
+
+    #[tokio::test]
+    async fn workspace_name_conflict_retries_with_fresh_suffix() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let workspace_name = super::deterministic_workspace_name();
+        let first_conflict = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/workspaces")
+                .json_body(json!({ "name": workspace_name }));
+            then.status(409)
+                .header("content-type", "application/json")
+                .body(
+                    r#"{"ok":false,"error":{"code":"workspace_already_exists","message":"Workspace already exists"}}"#,
+                );
+        });
+        let second_workspace = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/workspaces")
+                .body_contains(format!("\"name\":\"{workspace_name}-"));
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"workspace_id":"ws_fallback","api_key":"rk_live_fallback","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+        let register = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_fallback");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"id":"a_fallback","name":"lead","token":"at_live_fallback","status":"online","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+
+        let client = AuthClient::new(server.base_url());
+        let session = client.startup_session(Some("lead")).await.unwrap();
+
+        assert_eq!(session.token, "at_live_fallback");
+        assert_eq!(session.credentials.api_key, "rk_live_fallback");
+        assert_eq!(session.credentials.workspace_id, "ws_fallback");
+        first_conflict.assert_hits(1);
+        second_workspace.assert_hits(1);
+        register.assert_hits(1);
+    }
+
+    #[tokio::test]
+    async fn rotate_token_calls_rotate_endpoint_and_preserves_name() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let rotate = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents/lead/rotate-token")
+                .header("authorization", "Bearer rk_live_cached");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"token":"at_live_rotated","name":"lead"}}"#);
+        });
+
+        let client = AuthClient::new(server.base_url());
+
+        let cached = CredentialCache {
+            workspace_id: "ws_cached".into(),
+            workspace_alias: None,
+            agent_id: "a_old".into(),
+            api_key: "rk_live_cached".into(),
+            agent_name: Some("lead".into()),
+            agent_token: None,
+            updated_at: chrono::Utc::now(),
+        };
+
+        let session = client.rotate_token(&cached).await.unwrap();
+        assert_eq!(session.token, "at_live_rotated");
+        assert_eq!(session.credentials.agent_name.as_deref(), Some("lead"));
+        assert_eq!(session.credentials.agent_id, "a_old");
+        assert_eq!(session.credentials.workspace_id, "ws_cached");
+        rotate.assert_hits(1);
+    }
+
+    #[tokio::test]
+    async fn rotate_token_falls_back_to_reregister_on_404() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let rotate_404 = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents/lead/rotate-token")
+                .header("authorization", "Bearer rk_live_cached");
+            then.status(404)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":false,"error":{"code":"not_found","message":"not found"}}"#);
+        });
+        let register = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_cached");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"id":"a_new","name":"lead","token":"at_live_reregistered","status":"online","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+
+        let client = AuthClient::new(server.base_url());
+
+        let cached = CredentialCache {
+            workspace_id: "ws_cached".into(),
+            workspace_alias: None,
+            agent_id: "a_old".into(),
+            api_key: "rk_live_cached".into(),
+            agent_name: Some("lead".into()),
+            agent_token: None,
+            updated_at: chrono::Utc::now(),
+        };
+
+        let session = client.rotate_token(&cached).await.unwrap();
+        assert_eq!(session.token, "at_live_reregistered");
+        assert_eq!(session.credentials.agent_name.as_deref(), Some("lead"));
+        rotate_404.assert_hits(1);
+        register.assert_hits(1);
+    }
+
+    #[tokio::test]
+    async fn workspace_key_liveness_probe_returns_true_for_success() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let channels = server.mock(|when, then| {
+            when.method(GET)
+                .path("/v1/channels")
+                .header("authorization", "Bearer rk_live_cached");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":[]}"#);
+        });
+        let client = AuthClient::new(server.base_url());
+
+        let live = client
+            .workspace_key_is_live("rk_live_cached")
+            .await
+            .unwrap();
+        assert!(live);
+        channels.assert_hits(1);
+    }
+
+    #[tokio::test]
+    async fn workspace_key_liveness_probe_returns_false_for_unauthorized() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let channels = server.mock(|when, then| {
+            when.method(GET)
+                .path("/v1/channels")
+                .header("authorization", "Bearer rk_live_cached");
+            then.status(401)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":false,"error":{"code":"unauthorized","message":"unauthorized"}}"#);
+        });
+        let client = AuthClient::new(server.base_url());
+
+        let live = client
+            .workspace_key_is_live("rk_live_cached")
+            .await
+            .unwrap();
+        assert!(!live);
+        channels.assert_hits(1);
+    }
+
+    #[test]
+    fn deterministic_workspace_name_is_stable() {
+        let a = super::deterministic_workspace_name();
+        let b = super::deterministic_workspace_name();
+        assert_eq!(a, b);
+        assert!(a.starts_with("relay-"));
+        assert_eq!(a.len(), "relay-".len() + 8);
+    }
+
+    #[test]
+    fn workspace_key_normalization_accepts_rk_prefixes() {
+        assert_eq!(
+            super::normalize_workspace_key(" rk_test_123 "),
+            Some("rk_test_123".to_string())
+        );
+        assert_eq!(super::normalize_workspace_key("at_live_1"), None);
+    }
+
+    /// 403 Forbidden behaves identically to 401 Unauthorized — the env key
+    /// should be soft-rejected and a fresh workspace created.
+    #[tokio::test]
+    async fn forbidden_env_key_falls_back_to_fresh_workspace() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let _forbidden_register = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_forbidden");
+            then.status(403)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":false,"error":{"code":"forbidden","message":"forbidden"}}"#);
+        });
+        let workspace = server.mock(|when, then| {
+            when.method(POST).path("/v1/workspaces");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"workspace_id":"ws_fresh","api_key":"rk_live_fresh","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+        let fresh_register = server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_fresh");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"id":"a10","name":"broker","token":"at_live_10","status":"online","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+
+        unsafe {
+            std::env::set_var("RELAY_API_KEY", "rk_live_forbidden");
+        }
+
+        let client = AuthClient::new(server.base_url());
+        let session = client.startup_session(Some("broker")).await.unwrap();
+
+        // Must return the FRESH workspace key, not the forbidden env key.
+        assert_eq!(session.credentials.api_key, "rk_live_fresh");
+        assert_eq!(session.credentials.workspace_id, "ws_fresh");
+        assert_eq!(session.token, "at_live_10");
+        workspace.assert_hits(1);
+        fresh_register.assert_hits(1);
+
+        unsafe {
+            std::env::remove_var("RELAY_API_KEY");
+        }
+    }
+
+    /// When the env key is rejected and the broker falls back to a fresh
+    /// workspace, the session's api_key MUST be the fresh key — not the stale
+    /// env key. This is the exact scenario that caused workers to receive an
+    /// incorrect RELAY_API_KEY and fail MCP authentication.
+    #[tokio::test]
+    async fn session_api_key_is_never_the_rejected_env_key() {
+        let _env_guard = clear_relay_env();
+        let server = MockServer::start();
+        let stale_key = "rk_live_stale_must_not_appear_in_session";
+        server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", format!("Bearer {stale_key}"));
+            then.status(401)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":false,"error":{"code":"unauthorized","message":"unauthorized"}}"#);
+        });
+        server.mock(|when, then| {
+            when.method(POST).path("/v1/workspaces");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"workspace_id":"ws_ok","api_key":"rk_live_correct","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+        server.mock(|when, then| {
+            when.method(POST)
+                .path("/v1/agents")
+                .header("authorization", "Bearer rk_live_correct");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(r#"{"ok":true,"data":{"id":"a11","name":"broker","token":"at_live_11","status":"online","created_at":"2025-01-01T00:00:00Z"}}"#);
+        });
+
+        unsafe {
+            std::env::set_var("RELAY_API_KEY", stale_key);
+        }
+
+        let client = AuthClient::new(server.base_url());
+        let session = client.startup_session(Some("broker")).await.unwrap();
+
+        // The stale env key must NEVER appear in the returned session.
+        assert_ne!(
+            session.credentials.api_key, stale_key,
+            "session must not use the rejected env key — workers would get wrong RELAY_API_KEY"
+        );
+        assert_eq!(session.credentials.api_key, "rk_live_correct");
+
+        unsafe {
+            std::env::remove_var("RELAY_API_KEY");
+        }
+    }
+}
